requirements.txt
Flask==3.0.0
opencv-contrib-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
